Restore "Standalone" mode
"Standalone" mode used to be an alternate compile across the
entire codebase with a compile flag set but really no need for
that; the entire thing can be a runtime check and we can set the
standalone flag in the constructor of the processor now we
are over in juce 6 land, and it works! (It creates a few objects
in the plugin we may not need but they don't do anything).

Closes #9