Fix a mac drag problem related to trackpad
The "RMB Drag" gesture on a mac trackpad is two fingers down,
click, drag. Cool. But 2 fingers down can also begin a wheel event.
The mac trackpad is smart enough to not send any motion but it does
generate a zero distance wheel event which B-Step picks up and uses
to interfere with the drag event. So don't start the wheel timer on
a zero distance wheel event.

Closes #22